fix: various config fixes
