+++ lib/config.ts
@@ -1,10 +1,8 @@
 import { normalize } from "path";
-import { existsSync, readFileSync, writeFileSync } from "fs";
 import { prompt } from "enquirer";
-import { format, resolveConfig } from "prettier";
 import { CrawlConfig } from "./models";
-
-const REPO_CONFIG_PATH = "./deprecation-crawler.config.json";
+import { REPO_CONFIG_PATH } from "./constants";
+import { readFile, updateRepoConfig } from "./utils";
 
 export async function getConfig(): Promise<CrawlConfig> {
   const repoConfigFile = readFile(REPO_CONFIG_PATH) || "{}";
@@ -44,46 +42,19 @@ export async function getConfig(): Promise<CrawlConfig> {
         type: "input",
         name: "excludeGlob",
         message: `Glob to exclude files or folders (default ignores node_modules and path in .gitignore)`,
-        initial: repoConfig.excludeGlob,
+        initial: repoConfig.excludeGlob ? repoConfig.excludeGlob.join(" ") : "",
         skip: !!repoConfig.tsConfigPath || !!repoConfig.excludeGlob,
       },
     ]);
 
     userConfig.excludeGlob = globAnswer.excludeGlob
-      ? [
-          ...getDefaultGlob(),
-          ...globAnswer.excludeGlob.split(" ").map((p) => p.trim()),
-        ].filter(Boolean)
+      ? [...globAnswer.excludeGlob.split(" ").map((p) => p.trim())].filter(
+          Boolean
+        )
       : [];
   }
 
   const config = { groups: [], ...repoConfig, ...userConfig };
   updateRepoConfig(config);
   return config;
-}
-
-export function updateRepoConfig(config: CrawlConfig) {
-  const prettiedConfig = format(JSON.stringify(config), {
-    parser: "json",
-    ...resolveConfig.sync("./"),
-  });
-
-  writeFileSync(REPO_CONFIG_PATH, prettiedConfig);
-}
-
-function getDefaultGlob() {
-  let ignore = ["!./node_modules/**"];
-  const gitignore = readFile("./.gitignore");
-  ignore.push(
-    // remove empty entries and comments
-    ...gitignore.split(/\r?\n/).filter((p) => p && !p.startsWith("#"))
-  );
-  return ["./**/*{.ts,.js}", ...ignore];
-}
-
-function readFile(path: string) {
-  if (existsSync(path)) {
-    return readFileSync(path, "utf-8");
-  }
-  return "";
 }
