+++ lib/constants.ts
@@ -1,0 +1,1 @@
+export const REPO_CONFIG_PATH = "./deprecation-crawler.config.json";
